requirements.txt
langchain 
langchain-core
langchain-huggingface
python-dotenv
streamlit
transformers